requirements.txt
cherrypy==18.6.1
psycopg2==2.9.1
flask==2.0.1
mysql-connector-python==8.0.25
passlib==1.7.4